DONE: Added complete Vanilla Launcher Swapper script
Fixed the Use script extender launchers.sh. Just had to fix the LOADERLIBRARY variable to give the proper directory path. As well as declare the ATTRIBUTE var as a integar.

Completed the Swap-to-vanilla-launcher.sh added array lists for game ids, launchers, loaders, game names, and the install directory of the game to be called by index number during checksum.